pyproject.toml
[project]
name = "imagerater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu126" },
]
torchvision = [
    { index = "pytorch-cu126" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
